requirements.txt
libais
psycopg2
pandas
sqlalchemy
apache-airflow
interpret
setuptools
numpy
scipy
pyscaffold
python-dotenv
requests
scikit-learn